+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/stats/ModuleStats.kt
@@ -34,6 +34,7 @@ import foundry.gradle.namedLazy
 import foundry.gradle.properties.mapToBoolean
 import foundry.gradle.properties.setDisallowChanges
 import foundry.gradle.register
+import foundry.gradle.tasks.mustRunAfterSourceGeneratingTasks
 import foundry.gradle.topography.KnownFeatures
 import foundry.gradle.topography.ModuleTopography
 import foundry.gradle.topography.ModuleTopographyTask
@@ -62,9 +63,7 @@ import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.SkipWhenEmpty
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.TaskProvider
-import org.gradle.api.tasks.compile.JavaCompile
 import org.jetbrains.kotlin.gradle.internal.KaptTask
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompilationTask
 import org.jgrapht.alg.scoring.BetweennessCentrality
 import org.jgrapht.graph.DefaultEdge
 import org.jgrapht.graph.DirectedAcyclicGraph
@@ -169,18 +168,15 @@ public object ModuleStatsTasks {
 
     val generatedSourcesAdded = AtomicBoolean()
     val addGeneratedSources = {
-      val shouldConfigure =
-        locTask != null && generatedSourcesAdded.compareAndSet(false, true) && includeGenerated
-      if (shouldConfigure) {
-        locTask!!.configure {
-          generatedSrcsDir.setDisallowChanges(project.layout.buildDirectory.dir("generated"))
+      locTask?.let { locTask ->
+        val shouldConfigure = generatedSourcesAdded.compareAndSet(false, true) && includeGenerated
+        if (shouldConfigure) {
+          locTask.configure {
+            generatedSrcsDir.setDisallowChanges(project.layout.buildDirectory.dir("generated"))
+          }
+          locTask.mustRunAfterSourceGeneratingTasks(project)
         }
       }
-    }
-
-    linkToLocTask {
-      it.mustRunAfter(project.tasks.withType(JavaCompile::class.java))
-      it.mustRunAfter(project.tasks.withType(KotlinCompilationTask::class.java))
     }
 
     project.pluginManager.apply {